@raghav-stripe 2024-02-15
> I'm fine with this change, but then we need to change the docstring in `grpc.[With]WriteBufferSize` to remove the bit about allocating double for historical reasons.

Done!